[Scheduler] Delete old rAF implementation (#17252)
We've been using the message loop version for a while. Time to delete.
