+++ packages/scheduler/src/SchedulerFeatureFlags.js
@@ -8,5 +8,4 @@
 
 export const enableSchedulerDebugging = false;
 export const enableIsInputPending = false;
-export const enableMessageLoopImplementation = true;
 export const enableProfiling = __PROFILE__;
