+++ packages/scheduler/src/forks/SchedulerHostConfig.default.js
@@ -5,18 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {
-  enableIsInputPending,
-  enableMessageLoopImplementation,
-} from '../SchedulerFeatureFlags';
-
-// The DOM Scheduler implementation is similar to requestIdleCallback. It
-// works by scheduling a requestAnimationFrame, storing the time for the start
-// of the frame, then scheduling a postMessage which gets scheduled after paint.
-// Within the postMessage handler do as much work as possible until time + frame
-// rate. By separating the idle call into a separate event tick we ensure that
-// layout, paint and other browser work is counted against the available time.
-// The frame rate is dynamically adjusted.
+import {enableIsInputPending} from '../SchedulerFeatureFlags';
 
 export let requestHostCallback;
 export let cancelHostCallback;
@@ -83,10 +72,13 @@ if (
   const Date = window.Date;
   const setTimeout = window.setTimeout;
   const clearTimeout = window.clearTimeout;
-  const requestAnimationFrame = window.requestAnimationFrame;
-  const cancelAnimationFrame = window.cancelAnimationFrame;
 
   if (typeof console !== 'undefined') {
+    // TODO: Scheduler no longer requires these methods to be polyfilled. But
+    // maybe we want to continue warning if they don't exist, to preserve the
+    // option to rely on it in the future?
+    const requestAnimationFrame = window.requestAnimationFrame;
+    const cancelAnimationFrame = window.cancelAnimationFrame;
     // TODO: Remove fb.me link
     if (typeof requestAnimationFrame !== 'function') {
       console.error(
@@ -114,32 +106,20 @@ if (
     getCurrentTime = () => Date.now() - initialTime;
   }
 
-  let isRAFLoopRunning = false;
   let isMessageLoopRunning = false;
   let scheduledHostCallback = null;
-  let rAFTimeoutID = -1;
   let taskTimeoutID = -1;
 
-  let frameLength = enableMessageLoopImplementation
-    ? // We won't attempt to align with the vsync. Instead we'll yield multiple
-      // times per frame, often enough to keep it responsive even at really
-      // high frame rates > 120.
-      5
-    : // Use a heuristic to measure the frame rate and yield at the end of the
-      // frame. We start out assuming that we run at 30fps but then the
-      // heuristic tracking will adjust this value to a faster fps if we get
-      // more frequent animation frames.
-      33.33;
-
-  let prevRAFTime = -1;
-  let prevRAFInterval = -1;
-  let frameDeadline = 0;
-
-  let fpsLocked = false;
+  // Scheduler periodically yields in case there is other work on the main
+  // thread, like user events. By default, it yields multiple times per frame.
+  // It does not attempt to align with frame boundaries, since most tasks don't
+  // need to be frame aligned; for those that do, use requestAnimationFrame.
+  let yieldInterval = 5;
+  let deadline = 0;
 
   // TODO: Make this configurable
   // TODO: Adjust this based on priority?
-  let maxFrameLength = 300;
+  let maxYieldInterval = 300;
   let needsPaint = false;
 
   if (
@@ -151,13 +131,13 @@ if (
     const scheduling = navigator.scheduling;
     shouldYieldToHost = function() {
       const currentTime = getCurrentTime();
-      if (currentTime >= frameDeadline) {
-        // There's no time left in the frame. We may want to yield control of
-        // the main thread, so the browser can perform high priority tasks. The
-        // main ones are painting and user input. If there's a pending paint or
-        // a pending input, then we should yield. But if there's neither, then
-        // we can yield less often while remaining responsive. We'll eventually
-        // yield regardless, since there could be a pending paint that wasn't
+      if (currentTime >= deadline) {
+        // There's no time left. We may want to yield control of the main
+        // thread, so the browser can perform high priority tasks. The main ones
+        // are painting and user input. If there's a pending paint or a pending
+        // input, then we should yield. But if there's neither, then we can
+        // yield less often while remaining responsive. We'll eventually yield
+        // regardless, since there could be a pending paint that wasn't
         // accompanied by a call to `requestPaint`, or other main thread tasks
         // like network events.
         if (needsPaint || scheduling.isInputPending()) {
@@ -165,8 +145,8 @@ if (
           return true;
         }
         // There's no pending input. Only yield if we've reached the max
-        // frame length.
-        return currentTime >= frameDeadline + maxFrameLength;
+        // yield interval.
+        return currentTime >= maxYieldInterval;
       } else {
         // There's still time left in the frame.
         return false;
@@ -180,7 +160,7 @@ if (
     // `isInputPending` is not available. Since we have no way of knowing if
     // there's pending input, always yield at the end of the frame.
     shouldYieldToHost = function() {
-      return getCurrentTime() >= frameDeadline;
+      return getCurrentTime() >= deadline;
     };
 
     // Since we yield every frame regardless, `requestPaint` has no effect.
@@ -196,164 +176,57 @@ if (
       return;
     }
     if (fps > 0) {
-      frameLength = Math.floor(1000 / fps);
-      fpsLocked = true;
+      yieldInterval = Math.floor(1000 / fps);
     } else {
       // reset the framerate
-      frameLength = 33.33;
-      fpsLocked = false;
+      yieldInterval = 5;
     }
   };
 
   const performWorkUntilDeadline = () => {
-    if (enableMessageLoopImplementation) {
-      if (scheduledHostCallback !== null) {
-        const currentTime = getCurrentTime();
-        // Yield after `frameLength` ms, regardless of where we are in the vsync
-        // cycle. This means there's always time remaining at the beginning of
-        // the message event.
-        frameDeadline = currentTime + frameLength;
-        const hasTimeRemaining = true;
-        try {
-          const hasMoreWork = scheduledHostCallback(
-            hasTimeRemaining,
-            currentTime,
-          );
-          if (!hasMoreWork) {
-            isMessageLoopRunning = false;
-            scheduledHostCallback = null;
-          } else {
-            // If there's more work, schedule the next message event at the end
-            // of the preceding one.
-            port.postMessage(null);
-          }
-        } catch (error) {
-          // If a scheduler task throws, exit the current browser task so the
-          // error can be observed.
+    if (scheduledHostCallback !== null) {
+      const currentTime = getCurrentTime();
+      // Yield after `yieldInterval` ms, regardless of where we are in the vsync
+      // cycle. This means there's always time remaining at the beginning of
+      // the message event.
+      deadline = currentTime + yieldInterval;
+      const hasTimeRemaining = true;
+      try {
+        const hasMoreWork = scheduledHostCallback(
+          hasTimeRemaining,
+          currentTime,
+        );
+        if (!hasMoreWork) {
+          isMessageLoopRunning = false;
+          scheduledHostCallback = null;
+        } else {
+          // If there's more work, schedule the next message event at the end
+          // of the preceding one.
           port.postMessage(null);
-          throw error;
         }
-      } else {
-        isMessageLoopRunning = false;
+      } catch (error) {
+        // If a scheduler task throws, exit the current browser task so the
+        // error can be observed.
+        port.postMessage(null);
+        throw error;
       }
-      // Yielding to the browser will give it a chance to paint, so we can
-      // reset this.
-      needsPaint = false;
     } else {
-      if (scheduledHostCallback !== null) {
-        const currentTime = getCurrentTime();
-        const hasTimeRemaining = frameDeadline - currentTime > 0;
-        try {
-          const hasMoreWork = scheduledHostCallback(
-            hasTimeRemaining,
-            currentTime,
-          );
-          if (!hasMoreWork) {
-            scheduledHostCallback = null;
-          }
-        } catch (error) {
-          // If a scheduler task throws, exit the current browser task so the
-          // error can be observed, and post a new task as soon as possible
-          // so we can continue where we left off.
-          port.postMessage(null);
-          throw error;
-        }
-      }
-      // Yielding to the browser will give it a chance to paint, so we can
-      // reset this.
-      needsPaint = false;
+      isMessageLoopRunning = false;
     }
+    // Yielding to the browser will give it a chance to paint, so we can
+    // reset this.
+    needsPaint = false;
   };
 
   const channel = new MessageChannel();
   const port = channel.port2;
   channel.port1.onmessage = performWorkUntilDeadline;
 
-  const onAnimationFrame = rAFTime => {
-    if (scheduledHostCallback === null) {
-      // No scheduled work. Exit.
-      prevRAFTime = -1;
-      prevRAFInterval = -1;
-      isRAFLoopRunning = false;
-      return;
-    }
-
-    // Eagerly schedule the next animation callback at the beginning of the
-    // frame. If the scheduler queue is not empty at the end of the frame, it
-    // will continue flushing inside that callback. If the queue *is* empty,
-    // then it will exit immediately. Posting the callback at the start of the
-    // frame ensures it's fired within the earliest possible frame. If we
-    // waited until the end of the frame to post the callback, we risk the
-    // browser skipping a frame and not firing the callback until the frame
-    // after that.
-    isRAFLoopRunning = true;
-    requestAnimationFrame(nextRAFTime => {
-      clearTimeout(rAFTimeoutID);
-      onAnimationFrame(nextRAFTime);
-    });
-
-    // requestAnimationFrame is throttled when the tab is backgrounded. We
-    // don't want to stop working entirely. So we'll fallback to a timeout loop.
-    // TODO: Need a better heuristic for backgrounded work.
-    const onTimeout = () => {
-      frameDeadline = getCurrentTime() + frameLength / 2;
-      performWorkUntilDeadline();
-      rAFTimeoutID = setTimeout(onTimeout, frameLength * 3);
-    };
-    rAFTimeoutID = setTimeout(onTimeout, frameLength * 3);
-
-    if (
-      prevRAFTime !== -1 &&
-      // Make sure this rAF time is different from the previous one. This check
-      // could fail if two rAFs fire in the same frame.
-      rAFTime - prevRAFTime > 0.1
-    ) {
-      const rAFInterval = rAFTime - prevRAFTime;
-      if (!fpsLocked && prevRAFInterval !== -1) {
-        // We've observed two consecutive frame intervals. We'll use this to
-        // dynamically adjust the frame rate.
-        //
-        // If one frame goes long, then the next one can be short to catch up.
-        // If two frames are short in a row, then that's an indication that we
-        // actually have a higher frame rate than what we're currently
-        // optimizing. For example, if we're running on 120hz display or 90hz VR
-        // display. Take the max of the two in case one of them was an anomaly
-        // due to missed frame deadlines.
-        if (rAFInterval < frameLength && prevRAFInterval < frameLength) {
-          frameLength =
-            rAFInterval < prevRAFInterval ? prevRAFInterval : rAFInterval;
-          if (frameLength < 8.33) {
-            // Defensive coding. We don't support higher frame rates than 120hz.
-            // If the calculated frame length gets lower than 8, it is probably
-            // a bug.
-            frameLength = 8.33;
-          }
-        }
-      }
-      prevRAFInterval = rAFInterval;
-    }
-    prevRAFTime = rAFTime;
-    frameDeadline = rAFTime + frameLength;
-
-    // We use the postMessage trick to defer idle work until after the repaint.
-    port.postMessage(null);
-  };
-
   requestHostCallback = function(callback) {
     scheduledHostCallback = callback;
-    if (enableMessageLoopImplementation) {
-      if (!isMessageLoopRunning) {
-        isMessageLoopRunning = true;
-        port.postMessage(null);
-      }
-    } else {
-      if (!isRAFLoopRunning) {
-        // Start a rAF loop.
-        isRAFLoopRunning = true;
-        requestAnimationFrame(rAFTime => {
-          onAnimationFrame(rAFTime);
-        });
-      }
+    if (!isMessageLoopRunning) {
+      isMessageLoopRunning = true;
+      port.postMessage(null);
     }
   };
